openHarmony, few tweaks

+++ avalon/harmony/workio.py
@@ -32,31 +32,38 @@ def save_file(filepath):
     save_disabled = True
     temp_path = lib.get_local_harmony_path(filepath)
 
-    if os.path.exists(temp_path):
-        shutil.rmtree(temp_path)
+    if lib.server:
+        if os.path.exists(temp_path):
+            try:
+                shutil.rmtree(temp_path)
+            except Exception as e:
+                raise Exception(f"cannot delete {temp_path}") from e
 
-    lib.server.send(
-        {"function": "scene.saveAs", "args": [temp_path]}
-    )["result"]
+        lib.server.send(
+            {"function": "scene.saveAs", "args": [temp_path]}
+        )["result"]
 
-    lib.zip_and_move(temp_path, filepath)
+        lib.zip_and_move(temp_path, filepath)
 
-    lib.workfile_path = filepath
+        lib.workfile_path = filepath
 
-    func = """function add_path(path)
-    {
-        var app = QCoreApplication.instance();
-        app.watcher.addPath(path);
-    }
-    add_path
-    """
+        func = """function add_path(path)
+        {
+            var app = QCoreApplication.instance();
+            app.watcher.addPath(path);
+        }
+        add_path
+        """
 
-    scene_path = os.path.join(
-        temp_path, os.path.basename(temp_path) + ".xstage"
-    )
-    lib.server.send(
-        {"function": func, "args": [scene_path]}
-    )
+        scene_path = os.path.join(
+            temp_path, os.path.basename(temp_path) + ".xstage"
+        )
+        lib.server.send(
+            {"function": func, "args": [scene_path]}
+        )
+    else:
+        os.environ["HARMONY_NEW_WORKFILE_PATH"] = filepath.replace("\\", "/")
+
     save_disabled = False
